update newSys page and refactor
