+++ src/Components/SigDetailsTable/NewSigDetailsTable.js
@@ -22,7 +22,6 @@ import { SkeletonTable } from '@patternfly/react-component-groups';
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
 import TableToolbar from '@redhat-cloud-services/frontend-components/TableToolbar';
 import { DateFormat } from '@redhat-cloud-services/frontend-components/DateFormat';
-import { expandMatchMetadata } from '../Common';
 import {
   GET_HITS_TABLE,
   GET_SIGNATURE_DETAILS_TABLE,
@@ -34,7 +33,7 @@ import messages from '../../Messages';
 import EmptyAccount from '../SharedComponents/EmptyAccount';
 import './SigDetailsTable.scss';
 import { useWorkspaceFeatureFlag } from '../../Utilities/Hooks';
-import { BasicTable } from './helpers';
+import { BasicTable, generateCode } from '../helpers';
 
 const sortIndices = {
   1: 'DISPLAY_NAME',
@@ -213,28 +212,6 @@ const NewSigDetailsTable = ({ ruleName, affectedCount, isEmptyAccount }) => {
     const collapseRows = [...rows];
     const host = collapseRows[rowKey + 1].hostData;
     let hitlistData;
-
-    const generateCode = () => {
-      return host.matches
-        .map(
-          (match, key) => `Match Source: ${match.source}
-Offset: ${match.stringOffset}
-Match Data: ${match.stringData}
-Match Identifier: ${match.stringIdentifier}
-Match Scan Date: ${new Date(match.scanDate).toUTCString()}
-${
-  match.metadata
-    ? expandMatchMetadata(JSON.parse(match.metadata.replaceAll('\n', '')))
-    : ''
-}
-  ${
-    host.matches.length > 1 && key !== host.matches.length - 1
-      ? '~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~'
-      : ''
-  }`
-        )
-        .join('\n');
-    };
 
     collapseRows[rowKey] = { ...collapseRows[rowKey], isOpen };
