+++ src/Components/SysDetailsTable/NewSysDetailsTable.js
@@ -22,7 +22,6 @@ import TableToolbar from '@redhat-cloud-services/frontend-components/TableToolba
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/PrimaryToolbar';
 import CodeEditor from '../CodeEditor/CodeEditor';
 import { DateFormat } from '@redhat-cloud-services/frontend-components/DateFormat';
-import { expandMatchMetadata } from '../Common';
 import {
   GET_HITS_TABLE,
   GET_SYSTEMS_DETAILS_TABLE_PAGE,
@@ -31,6 +30,7 @@ import MessageState from '../MessageState/MessageState';
 import StatusLabel from '../StatusLabel/StatusLabel';
 import messages from '../../Messages';
 import { Icon } from '@patternfly/react-core';
+import { BasicTable, generateCode } from '../helpers';
 
 const sortIndices = {
   1: 'NAME',
@@ -154,45 +154,9 @@ const NewSysDetailsTable = ({ systemId }) => {
   const onCollapse = async (e, rowKey, isOpen) => {
     const collapseRows = [...rows];
     const host = collapseRows[rowKey + 1].hostData;
-    let status;
-
-    const generateCode = (loading, status) =>
-      host.matches
-        .map(
-          (match, key) =>
-            `Match Source: ${match.source}
-Match Status: ${loading ? 'Loading...' : status}
-Offset: ${match.stringOffset}
-Match Data: ${match.stringData}
-Match Identifier: ${match.stringIdentifier}
-Match Scan Date: ${new Date(match.scanDate).toUTCString()}
-${
-  match.metadata
-    ? expandMatchMetadata(JSON.parse(match.metadata.replaceAll('\n', '')))
-    : ''
-}
-${
-  host.matches.length > 1 && key !== host.matches.length - 1
-    ? '~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~'
-    : ''
-}`
-        )
-        .join('\n');
+    let hitlistData;
 
     collapseRows[rowKey] = { ...collapseRows[rowKey], isOpen };
-    collapseRows[rowKey + 1].cells = [
-      {
-        title: (
-          <CodeEditor
-            height="250px"
-            code={generateCode(true, '')}
-            isDownloadEnabled
-            isCopyEnabled
-          />
-        ),
-      },
-    ];
-    stateSet({ type: 'setRows', payload: collapseRows });
 
     try {
       const response = await client.query({
@@ -203,7 +167,7 @@ ${
         },
       });
 
-      status = response.data.hitsList[0].status;
+      hitlistData = response.data.hitsList;
     } catch (error) {
       console.error('Error fetching data:', error);
     }
@@ -211,12 +175,15 @@ ${
     collapseRows[rowKey + 1].cells = [
       {
         title: (
-          <CodeEditor
-            height="250px"
-            code={generateCode(false, status)}
-            isDownloadEnabled
-            isCopyEnabled
-          />
+          <div>
+            <BasicTable hitlistData={hitlistData} />
+            <CodeEditor
+              height="250px"
+              code={generateCode(host)}
+              isDownloadEnabled
+              isCopyEnabled
+            />
+          </div>
         ),
       },
     ];
