+++ src/Components/helpers.js
@@ -9,6 +9,7 @@ import {
   TextInputGroup,
   TextInputGroupMain,
 } from '@patternfly/react-core';
+import { expandMatchMetadata } from './Common';
 
 const filterValues = [
   { name: 'Not reviewed', value: 'NOT_REVIEWED' },
@@ -56,7 +57,6 @@ export const DropdownBasic = ({ initialValue }) => {
 
   return (
     <Dropdown
-      style={{ width: '200px' }}
       isOpen={isOpen}
       onSelect={onSelect}
       onOpenChange={(isOpen) => setIsOpen(isOpen)}
@@ -129,4 +129,26 @@ export const BasicTable = ({ hitlistData }) => {
 
 BasicTable.propTypes = {
   hitlistData: PropTypes.any,
+};
+
+export const generateCode = (host) => {
+  return host.matches
+    .map(
+      (match, key) => `Match Source: ${match.source}
+Offset: ${match.stringOffset}
+Match Data: ${match.stringData}
+Match Identifier: ${match.stringIdentifier}
+Match Scan Date: ${new Date(match.scanDate).toUTCString()}
+${
+  match.metadata
+    ? expandMatchMetadata(JSON.parse(match.metadata.replaceAll('\n', '')))
+    : ''
+}
+${
+  host.matches.length > 1 && key !== host.matches.length - 1
+    ? '~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~'
+    : ''
+}`
+    )
+    .join('\n');
 };
